# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for SmartLeague. Cookie-based auth (HttpOnly access/refresh cookies).",
        "title": "SmartLeague API",
        "contact": {
            "name": "SmartLeague Backend"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "paths": {
...
        "/api/v1/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginUserResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "user_auth_refresh_token=token; Path=/; HttpOnly; Secure; SameSite=Strict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Logout user",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "user_auth_refresh_token=; expires=Thu, 01 Jan 1970 00:00:00 GMT; Path=/; HttpOnly"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterUserResponse"
                        },
                        "headers": {
                            "Set-Cookie": {
                                "type": "string",
                                "description": "user_auth_refresh_token=token; Path=/; HttpOnly; Secure; SameSite=Strict"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "get": {
                "description": "Public profile view by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Update user by id. Only for admins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update each user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEachUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEachUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/{id}/series": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetUserSeriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.HTTPStatus"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GetUserResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.GetUserSeriesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PlayerSeries"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.PaginationInfo"
                }
            }
        },
        "dto.HTTPStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoginUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "minLength": 6,
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 100,
                    "minLength": 8,
                    "example": "SecurePass123!"
                }
            }
        },
        "dto.LoginUserResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.PlayerSeries": {
            "type": "object",
            "properties": {
                "end_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_at": {
                    "type": "string"
                }
            }
        },
        "dto.RegisterUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "minLength": 6,
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "mishmish"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 100,
                    "minLength": 8,
                    "example": "SecurePass123!"
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.RegisterUserResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.UpdateEachUserRequest": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "maximum": 3,
                    "minimum": 0,
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ]
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000,
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "maxLength": 254,
                    "minLength": 6,
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "mishmish"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "maxLength": 100,
                    "minLength": 8,
                    "example": "SecurePass123!"
                },
                "role": {
                    "$ref": "#/definitions/types.Role"
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.UpdateEachUserResponse": {
            "type": "object",
            "properties": {
                "club_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "club_state": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ClubState"
                        }
                    ],
                    "example": 0
                },
                "description": {
                    "type": "string",
                    "example": "About me"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "nickname": {
                    "type": "string",
                    "example": "mishmish"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Role"
                        }
                    ],
                    "example": 0
                },
                "show_name": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "types.ClubState": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "ClubStateNone",
                "ClubStateMember",
                "ClubStateLeader",
                "ClubStatePresident",
                "ClubStateResident"
            ]
        },
...
        "types.Role": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleModerator",
                "RoleAdmin",
                "RoleSuperAdmin"
            ]
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Authentication via HttpOnly cookies:\\n- `user_auth_access_token` (short-lived)\\n- `user_auth_refresh_token` (long-lived)\\n\\nProtected endpoints require valid cookies to be sent by the client.",
            "type": "apiKey",
            "name": "user_auth_access_token",
            "in": "cookie"
        }
    },
...
}